+++ package.json
@@ -18,6 +18,7 @@
     "@mapbox/rehype-prism": "^0.5.0",
     "@mdx-js/loader": "^1.6.6",
     "@next/mdx": "^9.4.4",
+    "glob": "^7.1.6",
     "next": "^9.4.4",
     "next-pwa": "^3.0.1",
     "normalize.css": "^8.0.1",
