+++ src/pages/index/index.tsx
@@ -1,19 +1,19 @@
-import { NextPage } from 'next'
+import { NextPage, GetStaticProps } from 'next'
+import Head from 'next/head'
 import React from 'react'
 import Layout from '~/components/layouts/Default'
-import { Metadata } from '*.mdx'
-import { getAllMetadata } from '~/utils/meta'
+import { getAllMetadata } from '~/utils/posts'
 import Post from '~/components/organisms/Post'
-import styles from './styles.module.scss'
-import Head from 'next/head'
 import OGP from '~/components/ogp'
+import { Metadata } from '*.mdx'
+import styles from './styles.module.scss'
 
 interface P {
-  metadata: Metadata[]
+  metaList: Metadata[]
 }
 
-const Page: NextPage<P> = ({ metadata }) => {
-  const latestMeta = metadata[0]
+const Page: NextPage<P> = ({ metaList }) => {
+  const latestMeta = metaList[0]
   const title = "yKicchan's blog"
   const description = `Web エンジニアが気ままにアウトプットしてる技術ブログ。\n${
     latestMeta.title
@@ -34,7 +34,7 @@ const Page: NextPage<P> = ({ metadata }) => {
       </Head>
       <Layout>
         <ul className={styles.posts}>
-          {metadata.map((meta) => (
+          {metaList.map((meta) => (
             <li key={meta.path}>
               <Post meta={meta} />
             </li>
@@ -45,11 +45,11 @@ const Page: NextPage<P> = ({ metadata }) => {
   )
 }
 
-export async function getStaticProps() {
-  const metadata = await getAllMetadata()
+export const getStaticProps: GetStaticProps<P> = async () => {
+  const metaList = await getAllMetadata()
   return {
     props: {
-      metadata,
+      metaList,
     },
   }
 }
