+++ src/utils/posts.ts
@@ -1,17 +1,21 @@
-import fs from 'fs'
 import { Metadata } from '*.mdx'
+import glob from 'glob'
 
 export async function getAllMetadata(): Promise<Metadata[]> {
-  const dir = '/src/pages/posts/'
-  const files = fs.readdirSync(process.cwd() + dir).reverse()
+  const pattern = 'src/pages/posts/**/*.mdx'
+  const files = glob.sync(pattern)
   const mdxDocs = Promise.all(
-    files.map((file) => import('~/pages/posts/' + file))
+    files.map((f) => import('../' + f.substring(f.indexOf('pages'))))
   )
   const metaList = (await mdxDocs).map((mdx, i) => ({
     ...mdx.meta,
-    path: `/posts/${subExt(files[i])}`,
+    path: toPostPath(files[i]),
   }))
   return metaList
+}
+
+function toPostPath(file: string) {
+  return subExt(file.substring(file.indexOf('/posts')))
 }
 
 function subExt(file: string) {
